refactor(parser-result): Split parser-result normalizers by language
- Split language-specific parser-result normalizers into dedicated files
- Rename combined helper files to match their remaining scopes
- Preserve normalization behavior while reducing file size and coupling
